development progress

+++ arduino/arduino.v1.ino
@@ -8,14 +8,14 @@
 #define TRIGGER_THRESHOLD_CM 13 // threshold for triggering "full" status
 #define TRIG_DURATION_US 10     // duration for triggering the ultrasonic pulse (in microseconds)
 
-const int TRIG_PINS[NUM_SENSORS] = {3, 5, 7, 9};
-const int ECHO_PINS[NUM_SENSORS] = {4, 6, 8, 10};
+const uint8_t TRIG_PINS[NUM_SENSORS] = {3, 5, 7, 9};
+const uint8_t ECHO_PINS[NUM_SENSORS] = {4, 6, 8, 10};
 float distances[NUM_SENSORS];
 
 void setup()
 {
     Serial.begin(115200);
-    for (int i = 0; i < NUM_SENSORS; i++)
+    for (uint8_t i = 0; i < NUM_SENSORS; i++)
     {
         pinMode(TRIG_PINS[i], OUTPUT);
         pinMode(ECHO_PINS[i], INPUT);
@@ -23,7 +23,7 @@ void setup()
     }
 }
 
-float getDistanceCM(int trigPin, int echoPin)
+float getDistanceCM(uint8_t trigPin, uint8_t echoPin)
 {
     digitalWrite(trigPin, LOW);
     delayMicroseconds(2);
@@ -40,7 +40,7 @@ float getDistanceCM(int trigPin, int echoPin)
 
 void readAllDistances()
 {
-    for (int i = 0; i < NUM_SENSORS; i++)
+    for (uint8_t i = 0; i < NUM_SENSORS; i++)
     {
         distances[i] = getDistanceCM(TRIG_PINS[i], ECHO_PINS[i]);
     }
